# Expanded for comprehsions with folding