--- a/tests/script_errors/bad_fold/script.tremor
+++ b/tests/script_errors/bad_fold/script.tremor
@@ -1,0 +1,3 @@
+for event of
+  case (k, v) => v
+into [] use merge end;