--- a/tests/script_runtime_errors/bad_fold_op/script.tremor
+++ b/tests/script_runtime_errors/bad_fold_op/script.tremor
@@ -1,0 +1,3 @@
+for event of
+  case (k, v) => v
+into [] use / end