--- a/tests/script_runtime_errors/bad_fold_op_record/script.tremor
+++ b/tests/script_runtime_errors/bad_fold_op_record/script.tremor
@@ -1,0 +1,3 @@
+for event of
+  case (k, v) => {"#{k}": v}
+into {} use / end