--- a/tests/script_runtime_errors/bad_fold_type_imut/script.tremor
+++ b/tests/script_runtime_errors/bad_fold_type_imut/script.tremor
@@ -1,0 +1,3 @@
+(for event of
+  case (k,v) => [k, v]
+into {} end)