--- a/tests/script_warnings/for_fold_op/script.tremor
+++ b/tests/script_warnings/for_fold_op/script.tremor
@@ -1,0 +1,3 @@
+for event of
+  case (k, v) => v
+into 0 use + end