--- a/tests/scripts/fold_array/script.tremor
+++ b/tests/scripts/fold_array/script.tremor
@@ -1,0 +1,3 @@
+for event of
+  case (k, v) => v + 1
+end;