--- a/tests/scripts/fold_array_imut/script.tremor
+++ b/tests/scripts/fold_array_imut/script.tremor
@@ -1,0 +1,3 @@
+(for event of
+  case (k, v) => v + 1
+end)