--- a/tests/scripts/fold_array_initial_op_imut/script.tremor
+++ b/tests/scripts/fold_array_initial_op_imut/script.tremor
@@ -1,0 +1,3 @@
+(for event of
+  case (k, v) => v + 1
+into [1] use + end)