… CHANGELOG.md
… static/language/epilog/for.md
… tests/script_errors/bad_fold/error.txt
… tests/script_errors/bad_fold_imut/error.txt
… tests/script_runtime_errors/bad_fold_op/error.txt
… tests/script_runtime_errors/bad_fold_op/in
… tests/script_runtime_errors/bad_fold_op_imut/error.txt
… tests/script_runtime_errors/bad_fold_op_imut/in
… tests/script_runtime_errors/bad_fold_op_record/error.txt
… tests/script_runtime_errors/bad_fold_op_record/in
… tests/script_runtime_errors/bad_fold_op_record_imut/error.txt
… tests/script_runtime_errors/bad_fold_op_record_imut/in
… tests/script_runtime_errors/bad_fold_type/error.txt
… tests/script_runtime_errors/bad_fold_type/in
… tests/script_runtime_errors/bad_fold_type_imut/error.txt
… tests/script_runtime_errors/bad_fold_type_imut/in
… tests/script_warnings/for_fold_op/warning.txt
… tests/scripts/fold_array/in
… tests/scripts/fold_array/out
… tests/scripts/fold_array_imut/in
… tests/scripts/fold_array_imut/out
… tests/scripts/fold_array_initial/in
… tests/scripts/fold_array_initial/out
… tests/scripts/fold_array_initial_imut/in
… tests/scripts/fold_array_initial_imut/out
… tests/scripts/fold_array_initial_op/in
… tests/scripts/fold_array_initial_op/out
… tests/scripts/fold_array_initial_op_imut/in
… tests/scripts/fold_array_initial_op_imut/out
… tests/scripts/fold_number/in