# Pull request

## Description

This PR expands `for` comprehension to allow for initial values and a fold operator. Both are optional and default, so the current behavior is maintained, and full backward compatibility is given.

The syntax changes from:

```
for <target> of
  <cases>
end;
```

to

```
for <target> of
  <cases>
[into <initial> [use <operator>]] end;
```

where `initial` defaults to `[]` and `operator` defaults to `+`.

Vaid operators are `+`, `-`, `*` and `/`.

The syntax leaves open the opportunity to expand `operator` to have more complex folding using lambdas or function pointers once those are implemented. 

## Checklist

* [x] The RFC, if required, has been submitted and approved
* [x] Any user-facing impact of the changes is reflected in docs.tremor.rs
* [x] The code is tested
* [x] Use of unsafe code is reasoned about in a comment
* [x] Update CHANGELOG.md appropriately, recording any changes, bug fixes, or other observable changes in behavior
* [x] The performance impact of the change is measured (see below)

## Performance

Summary: We gain nearly 20% performance for the record case while maintaining near parity / stay within the noise level for arrays.

Simple "for" loop creating an array

```tremor
    define script runtime
    script
      for event of
        case (k, v) => v
      end
    end;
```

Creating a record:

```tremor
# old
    define script runtime
    script
      use std::record;
      record::from_array(for event of
        case (k, v) => [k, v]
      end)
    end;
# new
    define script runtime
    script
      for event of
        case (k, v) => {"{k}": v}
      into {} end
    end;

```



| target | old | new| 
|-|-|-|
| array | 107.6 MB/s | 108.6 MB/s |
| record | 121.4 MB/s | 102.2 MB/s |