@Licenser 2023-09-18
I got to admit, I'm disappointed by the changes in performance, but it does read nicer at least