@mfelsche 2023-09-25
Would this also work as expected on strings with `+`?
I do think this table here: https://www.tremor.rs/docs/edge/language/expressions#operators needs some updating. And it should be referenced from the for-comprehension docs, in order to show what operations are allowed in the context of a for comprehension.