@darach 2023-09-25
I do not like this syntax at all and believe we need an RFC as head/tail pattern matching in functions would resolve this far more elegantly.